simple: extract SimpleConfig to separate package for gomobile bind
gomobile requires all exported types to be in directly bound packages.
Move SimpleConfig from runner to new simple package, add both mobile/
and simple/ to the gomobile bind command so all types are found.
